Bug Fixes and AddOns
The scheduler system can mapped the meteor script to the right filesystem now. It also handle local fs and hdfs and
use the right path seperator from the running system.

There are many bug fixes in that included. The most important one are:
- mapped the the meteor script right
- handle result requests and send file blocks in correct way (for local and high distributed filesystems)
- create directories for each client

IMPORTANT: There are integration tests in that commit. It'll be changed immediately!